Add support for external modules
Modules can now be run remotely, outside the Mir node,
in a dedicated module server.
The Mir node can use a proxy module that connects to this server
over a websocket interface and relays all the incoming events
to the remote server. The server applies the events to its local module
(which implements the actual module logic) and sends back the results.
Those resulting events are then emitted by the proxy module
(which lives in the Mir node).

Signed-off-by: Matej Pavlovic <matopavlovic@gmail.com>